[simulation/view] A different solution to the update problem
Let the tank tabs register the listener and have a attribute which
tracks wether the model update is because of the slider, to ignore it.
This means that the model recieves updates all the time, except when the
updates are by itself. This is different than before, because before the
window had only recieved updates, if the control elements were disabled.