[simulation/controller] Don't restart server if port didn't change
Don't do anything if two ports are the same. Write settings only if
restarting succeeds, because otherwise the port may be in the
settingsfile and starting the application fails. Also stop all servers
prior to restarting them, to make it possible to interchange ports of the
servers.
Also check if OPC UA server is null, because if restarting a server fails, it may be null.